% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civ.R
\name{summary.civ}
\alias{summary.civ}
\title{Inference Methods for the Categorical Instrumental Variable Estimator.}
\usage{
\method{summary}{civ}(object, ...)
}
\arguments{
\item{object}{An object of class \code{civ} as fitted by \code{\link[=civ]{civ()}}.}

\item{...}{Additional arguments passed to \code{summary.ivreg}. See
\code{\link[AER:summary.ivreg]{AER::summary.ivreg()}} for a complete list of arguments.}
}
\value{
An object of class \code{summary.ivreg} with inference results.
}
\description{
Inference methods for the categorical instrumental variable
estimators. Simple wrapper for \code{\link[AER:summary.ivreg]{AER::summary.ivreg()}}.
}
\examples{
# Use data from the included simulated dataset
y <- SimDat$y
D <- SimDat$D
Z <- SimDat$Z
X <- SimDat$X
# Estimate categorical instrument variable estimator with K = 3
civ_fit <- civ(y, D, Z, X, K = 3)
summary(civ_fit)
}
\references{
Fox J, Kleiber C, Zeileis A (2023). "ivreg: Instrumental-Variables Regression
by '2SLS', '2SM', or '2SMM', with Diagnostics". R package.

Wiemann T (2023). "Optimal Categorical Instruments."
}
\seealso{
\code{\link[AER:summary.ivreg]{AER::summary.ivreg()}}
}
